Break block dispatcher file up and move common code to base class
