+++ packages/node/src/indexer/blockDispatcher/index.ts
@@ -1,0 +1,12 @@
+// Copyright 2020-2021 OnFinality Limited authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { IBlockDispatcher } from './base-block-dispatcher';
+import { BlockDispatcherService } from './block-dispatcher.service';
+import { WorkerBlockDispatcherService } from './worker-block-dispatcher.service';
+
+export {
+  IBlockDispatcher,
+  BlockDispatcherService,
+  WorkerBlockDispatcherService,
+};
