+++ packages/node/src/indexer/fetch.service.test.ts
@@ -18,6 +18,7 @@ import { Sequelize } from 'sequelize';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import * as SubstrateUtil from '../utils/substrate';
 import { ApiService } from './api.service';
+import { BlockDispatcherService } from './blockDispatcher';
 import { DictionaryService } from './dictionary.service';
 import { DsProcessorService } from './ds-processor.service';
 import { DynamicDsService } from './dynamic-ds.service';
@@ -26,7 +27,6 @@ import { IndexerManager } from './indexer.manager';
 import { ProjectService } from './project.service';
 import { BlockContent } from './types';
 import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
-import { BlockDispatcherService } from './worker/block-dispatcher.service';
 
 const WS_ENDPOINT = 'wss://polkadot.api.onfinality.io/public-ws';
 const HTTP_ENDPOINT = 'https://polkadot.api.onfinality.io/public';
