+++ packages/node/src/indexer/fetch.service.ts
@@ -43,11 +43,11 @@ import * as SubstrateUtil from '../utils/substrate';
 import { calcInterval } from '../utils/substrate';
 import { yargsOptions } from '../yargs';
 import { ApiService } from './api.service';
+import { IBlockDispatcher } from './blockDispatcher';
 import { DictionaryService, SpecVersion } from './dictionary.service';
 import { DsProcessorService } from './ds-processor.service';
 import { DynamicDsService } from './dynamic-ds.service';
 import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
-import { IBlockDispatcher } from './worker/block-dispatcher.service';
 
 const logger = getLogger('fetch');
 let BLOCK_TIME_VARIANCE = 5000; //ms
